First pass semantic extensions 3.0 schema patch
This adds two patches to be applied to the standard 3.0 schema
to produce one with semantic annotations, plus a script to
apply them.

preliminary-patch.json is a JSON Patch document that moves certain
things from arrays to objects so that they can be affected by
JSON Merge Patch.  This is kept minimal as JSON Patch is much less
intuitive than JSON Merge Patch.

merge-patch.yaml is a JSON Merge Patch document (in YAML) that does
most of the work of adding the semantic annotations.

patch-schemas.py applies both patches, with minor additional tweaks
to keep things in a readable order.