--test-mode output, pointer template improvements
For system testing, we need predictable output, so omit triples
such as "locatedAt" that include local environment details.
Also always output nt11 for test mode, and serialize to a string
and sort the lines, removing blanks.  Return the resulting line
array, which should be comparable to the return from readlines()
of a stored test output file.

Re-open stdout (or stderr) in binary mode and serialize directly
to them to avoid holding the entire string in memory.  Return
the string if destination is None.

Improve [Relative] JSON Pointer Tempalte unit tests,
rename the module to a more accurate name, and fix some typos.