Significant rework of oasExamples & oasUniqueKey
Make both of these objects with enough info to validate
the conditions they imply:
    * relative location of objects across which keys must be unique
    * relative location of schema and encoding to validate examples

The Schema Object's "default" is also now treated as an example as
the validation requirements are the same.

Includes minor fixes to the schema patch that have come up.

Includes a workaround for refs into "/components/schemas"
during example validation, which will be removed once a larger
effort to teach jschon about OAS documents as mixutures of
schemas and non-schemas is complete.

Also use OAS-registered "commonmark" format instead of
"markdown-commonmark" which I'd made up for this project.

Set an order for annotation processing to ensure that references
are processed before validating examples.