Remove old prototype parser and early ideas
TODO: Decide if there's really a need to encapsulate the
JSON Schema implementation given the reliance on other jschon
features such as JSON Pointer support. For now, don't worry about
it but leave the base class in case it's worth revisiting.

Also remove the prototype that was used to pitch the project.