Rework schema loading and example validation
Add an OasJson class that can manage a mix of JSON Schema
and non-JSON Schema objects, enabled by the patches to
jschon previously added through pyproject.toml.

Add exceptions to report schema reference resolution problems,
some of which are caused by mis-configuration of URLs vs URIs,
and some of which are due to loading resources before there is
a way to determine if they are schemas or not.  This can be
fixed with a larger reworking, but for now require a command-line
argument to flag stand-along schema documents.

Add command-line options: -v for logging verbosity, and
-x for managing automatic suffix stripping for URIs vs URLs.