Generalize schema patching and support 3.1
This makes the schema patching more generic to support patching
and loading both the OAS 3.0 schema and the various OAS 3.1 schemas.

Ultimately, this infrastructure should be extensible for custom
schemas and schema patching.

The standard OAS 3.1 dialect metaschema has been modified to
require the format-assertion vocabulary and the extension vocabulary.
Techincally, the implementation provided for format-assertion is
still incomplete but it can easily be finished later by adding
more 3rd-party packages.

Add oasGroupType for inserting nodes that collect similar child
nodes that would otherwise be mixed with different node types.
This still requires additional work to support in the graph.

Usability fixes include writing out the partially patched
schema when there is an error so that the errors are more
easily debugged.

A parameter was added to the patch-oas-schemas command-line
tool to select the version, as 3.1 patches much more quickly
than 3.0 as it does not need to use alterschema first.