Do much less customizing/subclassing of jschon
Step 1: Disentangle loading from jschon subclasses.

Remove dependency on OASJSON and OASCatalog in favor of using
the new JSONFormat (or a relatively minimal subclass thereof).
Keep using the sources in oassource.py with the regular Catalog.

With that done, remove the oasjson and oascatalog modules entirely,
migrating the few bits that need to be kept elsewhere such as oas3dialect.

jschon has changed enough that it would be better to re-add new
versions of these if needed than try to integrate the new changes,
which were designed to make some of the things done here unnecessary,
into them.

Remove custom URI/IRI/[Relative]JSONPointer module, as having it
and retaining interoperability with jschon was challenging, and
jschon has agreed to at some point take on a different URI library
which will solve the underlying problems properly.

Step 2: Add OASResourceManager for OASSource management

Rather than subclassing jschon.Catalog, which gets complicated,
this class is essentially a proxy wrapping it to pipe the
source loading data between the OASScource classes and the
rest of oascomply.

With the proposed jschon extensibility changes, subclassing
is (probably, hopefully) no longer necessary.  These source
loading data communication channels may get pushed up to
jschon.Catalog later or they may not, but either way this should
continue to work.

Includes a placeholder oasTypeGroup as oasType implementation.
This needs to be replaced with a proper treatment of oasTypeGroup
once its nature and usage is settled.