Reorganize and unit test CLI code
Removed outdated things, settled on URI / URL vs Uri / Url,
broke up load() into testable functions.

Add .coveragerc to support dynamic contexts, reach
100% coverage of ThingToURI classes and parse_logging.

Numerous smaller fixes:

pathlib.Path.as_uri() correctly does not append a trailing '/'
for directories, but we need to do that for the expected behavior
with URI prefixes mapping to directories in jschon.

Check URL prefixes for trailing slash.

Remove never-implemented "store" option, which was for an
idea that is not urgent.  In the future, support for connecting
to another RDF database will be useful for supporting large
graphs, but that is best added when needed.  Remove the option
because it's confusing to have unimplemented options.

Implement revised --initial option using URIs to make it
easier to treat it like a "$ref" into the first document
instead of requiring custom handling.