Create and partially test OASNode class tree
OASNodeBase is a mixin to give a shared base to
various document node classes that hook into the jschon
document hierarchy at differentp places.  It is mostly
there to share the oasversion logic.

OASFormat includes complete documents (OASDocument)
and fragments embedded in other documents (OASFragment).

OASNode is for nodes within an OAS structure that do not
need special functionality.  It is not a format as it
does not know its metadocument (schema), as tracking
that through an entire OAS structure would be complex
and would not add value when what we care about is
the graph in the end.

OASContainer is a separate non-format class for the
documents containing OASFragments.