Many fixes to schema patching, esp for 3.1
Add intermediateRelation to oasIntermediate

json schema spec submodule for patching the std metaschemas

Fix 3.1 schema patching, add patched 2020-12 dialect.

Note that we rely on format switch detection to find the right
metaschema for the OAS 3.1 Schema Object, so there is no need
to anchor or reference anything in our version of the main
OAS 3.1 schema.