Use validate() with the metadocument
Shift to using validate(), which understands metadocument
switching (e.g. changing "$schema" in an embedded resource,
but also respecting "jsonSchemaDialect" without needing to
construct a different overall schema each time).

For the moment, this involves an egregious hack to support
determining the relevant "jsonSchemaDialect", which will
need to be revisited (and may or may not cover all cases,
due to the ambiguity of referencing into a document without
parsing the whole thing).